+++ CHANGELOG.md
@@ -10,3 +10,6 @@ All notable changes to Dexter will be documented in this file.
 - Liquidity pool fee fix for SundaeSwap when constructing pools from on-chain data. 
 - Add `SwapRequest.withSwapOutAmount(bigint)` to calculate the estimated swap in amount.
 - Add `SwapRequest.withSwapOutToken(Token)` to allow crafting a SwapRequest given the swap out token.
+- Update `FetchRequest.forDexs()` to `FetchRequest.onDexs()`.
+- Update `FetchRequest.forAllDexs()` to `FetchRequest.onAllDexs()`.
+- Add `FetchRequest.forTokens()` & `FetchRequest.forTokenPairs()` for filtering pools containing tokens/token pairs.